requirements.txt
# Database Drivers
psycopg2-binary>=2.9.9
pymysql>=1.1.0
sqlalchemy>=2.0.23

# Graph Databases
neo4j>=5.15.0
python-arango>=7.6.0

# Benchmarking and Utilities
pandas>=2.1.3
numpy>=1.26.2
matplotlib>=3.8.2
seaborn>=0.13.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-benchmark>=4.0.0

